Добавляет сброс слайдера при отправке и ресета формы
